Concluding authorization
