+++ server/route/booksRoute.js
@@ -2,34 +2,11 @@ const express = require('express');
 const router = express.Router();
 const booksService = require('../service/booksService');
 const authService = require('../service/authService');
-
-const SECRET = 'api_books';
+const auth = require('../middleware/auth');
 
 const jwt = require('jsonwebtoken');
 
-router.post('/login', async function (req, res, next) {
-    const user = req.body;
-
-    try {
-        existsUser = await authService.login(user);
-        const token = jwt.sign({ id: existsUser.id }, SECRET, { expiresIn: 60 });
-        res.status(200).json({ auth: true, token });
-    } catch (e) {
-        next(e);
-    }
-});
-
-function verifyJWT (req,res, next) {
-    const token = req.headers['x-access-token'];
-    jwt.verify(token, SECRET, (err, decoded) => {
-        if(err) return res.status(401).json({ auth: false});
-
-        req.id = decoded.id;
-        next();
-    });
-}
-
-router.get('/books', verifyJWT, async function (req, res,next) {
+router.get('/books', auth, async function (req, res,next) {
     try{
         const books = await booksService.getBooks();
         res.status(200).json(books);
@@ -38,7 +15,7 @@ router.get('/books', verifyJWT, async function (req, res,next) {
     }
 });
 
-router.get('/books/:id', verifyJWT, async function (req, res,next) {
+router.get('/books/:id', auth, async function (req, res,next) {
     const idBook = req.params.id;
     try {
         const book = await booksService.getBook(idBook);
@@ -48,7 +25,7 @@ router.get('/books/:id', verifyJWT, async function (req, res,next) {
     }
 });
 
-router.post('/books', verifyJWT, async function (req, res,next) {
+router.post('/books', auth, async function (req, res,next) {
     const book = req.body;
     try {
         const newBook = await booksService.saveBook(book);
@@ -58,7 +35,7 @@ router.post('/books', verifyJWT, async function (req, res,next) {
     }
 });
 
-router.put('/books/:id', verifyJWT, async function (req, res,next) {
+router.put('/books/:id', auth, async function (req, res,next) {
     const book = req.body;
     try {
         await booksService.updateBook(req.params.id, book);
@@ -68,7 +45,7 @@ router.put('/books/:id', verifyJWT, async function (req, res,next) {
     }
 });
 
-router.patch('/books/:id', verifyJWT, async function (req, res,next) {
+router.patch('/books/:id', auth, async function (req, res,next) {
     const book = req.body;
     try {
         await booksService.updateRentedBook(req.params.id, book);
@@ -78,7 +55,7 @@ router.patch('/books/:id', verifyJWT, async function (req, res,next) {
     }
 });
 
-router.delete('/books/:id', verifyJWT, async function (req, res,next) {
+router.delete('/books/:id', auth, async function (req, res,next) {
     try {
         await booksService.deleteBook(req.params.id);
         res.status(204).end();
